+++ ValladoCalc/BusinessLogic/ValladoCalc.BusinessLogic.Models/ExportModels/RV2COEResultModel.cs
@@ -2,16 +2,16 @@
 {
     public class RV2COEResultModel
     {
-        public decimal SemiParameter { get; set; }
-        public decimal SemiMajorAxis { get; set; }
-        public decimal Eccentricity { get; set; }
-        public decimal Inclination { get; set; }
-        public decimal AscendingNode { get; set; }
-        public decimal ArgumentOfPerigee { get; set; }
-        public decimal LongituteOfPerigee { get; set; }
-        public decimal TrueAnomaly { get; set; }
-        public decimal ArgumentOfLatitude { get; set; }
-        public decimal TrueLongitude { get; set; }
-        public decimal TrueLongitudeOfPerigee { get; set; }
+        public double SemiParameter { get; set; }
+        public double SemiMajorAxis { get; set; }
+        public double Eccentricity { get; set; }
+        public double Inclination { get; set; }
+        public double AscendingNode { get; set; }
+        public double ArgumentOfPerigee { get; set; }
+        public double LongituteOfPerigee { get; set; }
+        public double TrueAnomaly { get; set; }
+        public double ArgumentOfLatitude { get; set; }
+        public double TrueLongitude { get; set; }
+        public double TrueLongitudeOfPerigee { get; set; }
     }
 }
