+++ ValladoCalc/BusinessLogic/ValladoCalc.BusinessLogic.Services/Interfaces/Services/IDateTimeService.cs
@@ -2,6 +2,6 @@ namespace ValladoCalc.BusinessLogic.Services.Interfaces.Services
 {
     public interface IDateTimeService
     {
-        public Task<decimal> CalculateJulianDate(DateTime dateTime);
+        public Task<double> CalculateJulianDate(DateTime dateTime);
     }
 }
